better casts

+++ include/nil/blueprint/components/algebra/fixedpoint/plonk/cmp.hpp
@@ -274,7 +274,7 @@ namespace nil {
                 const typename plonk_fixedpoint_cmp<BlueprintFieldType, ArithmetizationParams>::input_type
                     &instance_input) {
 
-                int64_t start_row_index = 1 - component.rows_amount;
+                int64_t start_row_index = 1 - static_cast<int64_t>(component.rows_amount);
                 const auto var_pos = component.get_var_pos(static_cast<int64_t>(start_row_index));
 
                 using var = typename plonk_fixedpoint_cmp<BlueprintFieldType, ArithmetizationParams>::var;
